+++ src/lib/bitstream.h
@@ -31,13 +31,13 @@ extern "C"
      * Copies at most the specified number of bytes to 'dest', returning the number of bytes that were actually copied.
      * The data in the stream must be byte aligned, otherwise the behavior is undefined.
      */
-    int bitstream_copy_bytes(bitstream* stream, int bytesToRead, uint8_t* dest);
+    uint16_t bitstream_copy_bytes(bitstream* stream, uint16_t bytesToRead, uint8_t* dest);
 
     /*
      * Reads the specified number of bits, writing to 'result'. This function returns 1 if all bits could be read and 0
      * if more data is needed. In the failure case, the contents of 'result' are unspecified and no data is consumed
      */
-    int bitstream_read_bits(bitstream* stream, int bitsToRead, uint16_t* result);
+    int bitstream_read_bits(bitstream* stream, uint8_t bitsToRead, uint16_t* result);
 
     /*
      * Peeks whatever data is available, returning the number of bits in the result. The data is NOT consumed
@@ -52,7 +52,7 @@ extern "C"
     void bitstream_consume_bits(bitstream* stream, int bits);
 
     /* Same as the above functions, but does not check to verify that the bitstream has enough input data */
-    uint16_t bitstream_read_bits_unchecked(bitstream* stream, int bitsToRead);
+    uint16_t bitstream_read_bits_unchecked(bitstream* stream, uint8_t bitsToRead);
     uint16_t bitstream_peek_unchecked(bitstream* stream);
 
 #ifdef __cplusplus
